effects of introducing `_r8` and `implicit none` to code
The notebooks shows the effect of introducing `_8` to all
literal real values and replacing all `implicit real`
statements with `implicit none`.  The introduction of these
are done in stages.

1. _8 is appended to all literal real values in the code:
   the main program and all subroutines.
2. `implicit none` placed in the subroutine transf()
3. `implicit none` placed in the subroutine irflux()
4. `implicit none` placed in the subroutine voigh2()
5. `implicit none` placed in the function vibpf()
6. `implicit none` placed in the subroutine lblabs()

In this commit, comparison of results (in this case
 the absorption coefficient) shows that a difference
is produced in going from 4 to 5, and in going from
5 to 6, which is significant.

In addition are shown some results related to the investigation
of the difference arising from going from 5 to 6.