Update method for displaying results
CLIRAD's results are now loaded from only 2 files.  Comparison
of flux or cooling rate values between different models are
done by first picking a model as the benchmark and
the layers/levels to compare, and then look for the nearest
layers/levels in other models.  When taking the difference
between different models, because the layers/levels (their pressures)
may not be identical, to force them to be identical so that
Pandas will subtract the corresponding values and not give NaN,
the indices of other models are replaced with the benchmark's.