O3 band05 mls75. 20 ngs. (Pref, Tref)=(50,250)
dv = 0.005

Atmosphere profile: midlatitude summer
tsfc = 294 K

k-distribution method:
absths = min_2e-22_max_5e-17_ng_20_dlogN_uniform
wgt = all 1

crd: cc55ab5 (branch: lblnew_use_absmean)

Comments:
It is suspected that the previous Pref of 50 mb might be
too high and there might still be a range of pressure
between the lowest pressure in a typical atmosphere profile
and it that is still quite Lorentzian.  When absorption line
profiles are very much Lorentzian, the line half width
is proportional to the pressure.  This means that at higher pressures
the absorption coefficient have a narrower range, while
at lower pressure there is a larger range.

In order for the g-groups derived at Pref to be generally applicable,
the range of absorption coefficient at Pref ought to be
wider than that at all other pressures in the atmosphere profile.
Hence, by decreasing the Pref from 50 mb to 1 mb, we have
increased the range of absorption coefficients, and we hope
that this will mean that the g-groups derived from this
will apply better to the pressures in the atmosphere profiles.

The typical atmosphere profile will have pressures lower than even
1 mb.  But it is likely that, at these pressures, the line profile
have become Doppler.  In this case, the half-width are no longer
proportional to P, and hence the range of absorption coefficient
 no longer increases linearly with P.