Sort wavenumers by k at 2 (P, T)s. Toy k.
This notebook simulates the sorting of
wavenumbers into g-groups by using the
k distribution at two (pressure, temperature)
pairs.  The example uses very simple
k distriutions.  There are also plots that
try to show what is going on with this method
of sorting.

The upper limit of the highest g-group in the
low-reference section is determined through
the conservation of the total number of spectral points
using the distribution of k at all spectral points at
the low reference.