Sort wavenumbers in O3 band 5. Two references.
Sort the spectral points in O3 band 5 into
g-groups.  The spectral points in O3's
band 5 (980 cm-1 ~ 1100 cm-1) are binned by
using k(pt = (1 mb, 250 K)) for the higher g's,
and by using k(pt = (50 mb, 250 K)) for the lower
g's.

The bin edges for the lower g-groups are determined
by finding the range of the lowest percentile
occupied by the spectral points that are left
over by the sorting for the higher g's.