Sort wavenumbers in O3 band 5. Two references.
Sort the spectral points in O3 band 5 into
g-groups.  The spectral points in O3's
band 5 (980 cm-1 ~ 1100 cm-1) are binned by
using k(pt = (1 mb, 250 K)) for the higher g's,
and by using k(pt = (50 mb, 250 K)) for the lower
g's.

dv = .002 cm-1

The upper limit of the highest bin
for the lower g-groups is maximum of
k(pt = (50 mb, 250 K)) at the leftover spectral points.
The lower limit of the lowest bin for
the lower g-groups is the value previously selected
such that the k values below it can be averaged
using a linear mean.

The k-distribution function obtained here is equal
to that produced by lblnew.f, so this Python
implementation of the sorting method in lblnew.f
can be said to have passed testing.