Create pipeline for submitting lblnew.f runs
The pipeline aims to automate the following:

(1) Create directory in which to run an example/case
    (call it the 'case directory')
(2) Copy the source code from the `crd` git
    repository to the case directory.
(3) Enter the input parameters into 'lblnew.f'.
(4) Compile 'lblnew.f'.
(5) Execute the compiled output as a Python subprocess
(6) Move onto the next user-specified case, if given,
     and repeat.