Nbcovnert execute with no timeout
I think that in executing many notebooks at the same time,
because they are all trying to access IPython's history
database, the database becomes locked for some notebooks.
For these notebooks, because they have to wait for the history
database to become unlocked for them, the execution gets timed
out.

The 2 issues are related to the errors seen:

https://github.com/jupyter/nbconvert/issues/277
https://github.com/ipython/ipython/issues/2845

The following changes seem to make all the notebooks finish
execution successfully:

1. Use 'jupyter nbconvert --ExecutePreprocessor.timeout=None ...'
2. Update jupyter. `conda update conda`, then
   `conda update jupyter`
   `conda update ipython`